+++ crates/bevy_ui/src/render/queue.rs
@@ -1,0 +1,97 @@
+use crate::render::UiBatch;
+use crate::render::UiImageBindGroups;
+use crate::render::UiMeta;
+use crate::DrawUi;
+use crate::UiPipeline;
+use crate::{TransparentUi, UiPipelineKey};
+use bevy_asset::AssetEvent;
+use bevy_ecs::prelude::*;
+use bevy_render::view::Msaa;
+use bevy_render::{
+    render_asset::RenderAssets,
+    render_phase::{DrawFunctions, RenderPhase},
+    render_resource::*,
+    renderer::RenderDevice,
+    texture::Image,
+    view::{ExtractedView, ViewUniforms},
+};
+use bevy_sprite::SpriteAssetEvents;
+#[cfg(feature = "bevy_text")]
+use bevy_utils::FloatOrd;
+
+#[allow(clippy::too_many_arguments)]
+pub fn queue_uinodes(
+    draw_functions: Res<DrawFunctions<TransparentUi>>,
+    render_device: Res<RenderDevice>,
+    mut ui_meta: ResMut<UiMeta>,
+    view_uniforms: Res<ViewUniforms>,
+    ui_pipeline: Res<UiPipeline>,
+    mut pipelines: ResMut<SpecializedRenderPipelines<UiPipeline>>,
+    pipeline_cache: Res<PipelineCache>,
+    mut image_bind_groups: ResMut<UiImageBindGroups>,
+    gpu_images: Res<RenderAssets<Image>>,
+    ui_batches: Query<(Entity, &UiBatch)>,
+    mut views: Query<(&ExtractedView, &mut RenderPhase<TransparentUi>)>,
+    events: Res<SpriteAssetEvents>,
+    msaa: Res<Msaa>,
+) {
+    // If an image has changed, the GpuImage has (probably) changed
+    for event in &events.images {
+        match event {
+            AssetEvent::Created { .. } => None,
+            AssetEvent::Modified { handle } | AssetEvent::Removed { handle } => {
+                image_bind_groups.values.remove(handle)
+            }
+        };
+    }
+
+    if let Some(view_binding) = view_uniforms.uniforms.binding() {
+        ui_meta.view_bind_group = Some(render_device.create_bind_group(&BindGroupDescriptor {
+            entries: &[BindGroupEntry {
+                binding: 0,
+                resource: view_binding,
+            }],
+            label: Some("ui_view_bind_group"),
+            layout: &ui_pipeline.view_layout,
+        }));
+        let draw_ui_function = draw_functions.read().id::<DrawUi>();
+        for (view, mut transparent_phase) in &mut views {
+            let pipeline = pipelines.specialize(
+                &pipeline_cache,
+                &ui_pipeline,
+                UiPipelineKey {
+                    hdr: view.hdr,
+                    msaa_samples: msaa.samples(),
+                },
+            );
+            for (entity, batch) in &ui_batches {
+                image_bind_groups
+                    .values
+                    .entry(batch.image.clone_weak())
+                    .or_insert_with(|| {
+                        let gpu_image = gpu_images.get(&batch.image).unwrap();
+                        render_device.create_bind_group(&BindGroupDescriptor {
+                            entries: &[
+                                BindGroupEntry {
+                                    binding: 0,
+                                    resource: BindingResource::TextureView(&gpu_image.texture_view),
+                                },
+                                BindGroupEntry {
+                                    binding: 1,
+                                    resource: BindingResource::Sampler(&gpu_image.sampler),
+                                },
+                            ],
+                            label: Some("ui_material_bind_group"),
+                            layout: &ui_pipeline.image_layout,
+                        })
+                    });
+                transparent_phase.add(TransparentUi {
+                    draw_function: draw_ui_function,
+                    pipeline,
+                    entity,
+                    sort_key: FloatOrd(batch.z),
+                });
+            }
+        }
+    }
+}
